requirements.txt
python-daemon
dnspod
